Adding online check in addd list

+++ SmartVocalListAndroidProject/app/src/main/java/com/example/ronen/smartvocallist/Controller/AddListActivity.java
@@ -1,6 +1,7 @@
 package com.example.ronen.smartvocallist.Controller;
 
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.app.AppComponentFactory;
 import android.content.Intent;
 import android.graphics.Bitmap;
@@ -8,6 +9,7 @@ import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
 
+import android.os.StrictMode;
 import android.provider.MediaStore;
 import android.view.Gravity;
 import android.view.View;
@@ -88,7 +90,19 @@ public class AddListActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         setContentView(R.layout.activity_add_list);
         super.onCreate(savedInstanceState);
+
         model = ViewModelProviders.of(this).get(AddListViewModel.class);
+
+
+
+        if (!isInternetAvailable())
+        {
+            AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+            alertDialogBuilder.setMessage("Check internet connection if you are offline your acctions won't be saved");
+            alertDialogBuilder.setCancelable(true);
+            AlertDialog alertDialog = alertDialogBuilder.create();
+            alertDialog.show();
+        }
 
 
         ll = (LinearLayout) findViewById(R.id.llItem);
@@ -139,7 +153,9 @@ public class AddListActivity extends AppCompatActivity {
 
                 if (model.isUpdate)
                 {
-                    mod.deleteItem(model.Old);
+                    if (isInternetAvailable()) {
+                        mod.deleteItem(model.Old);
+                    }
                 }
 
  		        if(model.bitmap != null){
@@ -147,7 +163,9 @@ public class AddListActivity extends AppCompatActivity {
                         @Override
                         public void onComplete(String url) {
                             model.getCurrentChecklist().setUrl(url);
-                            mod.addItem(model.getCurrentChecklist());
+                            if (isInternetAvailable()) {
+                                mod.addItem(model.getCurrentChecklist());
+                            }
                             savingProgressBar.setVisibility(View.INVISIBLE);
                             finish();
                         }
@@ -155,7 +173,9 @@ public class AddListActivity extends AppCompatActivity {
                 }
                 else{
                     model.getCurrentChecklist().setUrl(model.URL);
-                    mod.addItem(model.getCurrentChecklist());
+                    if (isInternetAvailable()) {
+                        mod.addItem(model.getCurrentChecklist());
+                    }
                     savingProgressBar.setVisibility(View.INVISIBLE);
                     finish();
                 }
@@ -295,9 +315,16 @@ public class AddListActivity extends AppCompatActivity {
 
     private boolean isInternetAvailable() {
         try {
+            StrictMode.ThreadPolicy policy = new StrictMode.ThreadPolicy.Builder().permitAll().build();
+            StrictMode.ThreadPolicy safePolicy = StrictMode.getThreadPolicy();
+            StrictMode.setThreadPolicy(policy);
+
             InetAddress ipAddr = InetAddress.getByName("google.com");
+
+            StrictMode.setThreadPolicy(safePolicy);
             //You can replace it with your name
             return !ipAddr.equals("");
+
 
         } catch (Exception e) {
             return false;
